Refactor. Code moved to lib.rs.
